Add TransportNeeds
Add a model named TransportNeed which belongs_to a Person and a Conference.
This model stores a datetime field, a note text and a boolean flag
indicating whether the transport has been booked yet.

This is necessary in order to keep track of shuttles needed to bring
speakers and other persons to the venue and back.

It also adds a t_boolean helper that pretty-prints boolean values using
the simple_form translations. This may be useful for other stuff as well.