Ticket: make association polymorphic
Allow tickets to belong to other models than events. For this to work, the
'event_id' field is renamed to 'object_id', and a new column named
'object_type' is added, which is initially set to 'Event' during the
migration so all existing remote tickets keep working.

The TicketController is prepared for associtations of other types, but this
commit does not introduce any of them, and shouldn't result in any visible
change.