Person: add ticket association
Now that the ticket model is prepared for polymorphic associations, it can
be attached to other models easily.

This patch allows users to create tickets for persons. The UI is similar to
that of events: a button can be used to create a ticket, the _table and
_show partials show a link to the external ticket system, and editing is
allowed through _form.