event person: add filter input for person list
When adding people to an event, people complained the fact that the list is
only ordered by the full name of persons. If the user only knows the email
address of a person, for instance, there is no good way to find that entry.

This patch introduces a new text input to the event person input fields
which is unrelated to the model but only serves as helper during editing.
When the value changes, the select field is emptied and then filled again
with records that match the query. If the filter field is reset to empty,
all options are shown again. This way, it is possible to filter persons by
their first name, last name, email address or user ID.

A callback on cocoon:after-insert is added so the logic described above also
applied to dynamically appended entries.